apps_reworked/forensic_file_analyzer/requirements.txt
fastapi
uvicorn
pywebview
pefile
python-magic
python-multipart
